requirements.txt
wolframalpha
fastapi
requests
pillow
numpy
python-multipart